[Flags] Split converter into separate file
