+++ flags/flags.py
@@ -1,89 +1,19 @@
 import contextlib
 import datetime
 import discord
-import pycountry
 
 try:
     import tabulate
 except ModuleNotFoundError:
     tabulate = None
-    
-from typing import Dict, Union
 
 from redbot.core import commands
-from redbot.core.commands import BadArgument, Cog, Context, Converter
+from redbot.core.commands import Cog, Context
 from redbot.core.utils.chat_formatting import box, humanize_list
 
+from .functions import format_attr
+from .converters import CountryConverter
 
-def square(t):
-    return f"[{t}]"
-
-
-def format_attr(t):
-    return t.replace('_',' ').title()
-
-
-EXCEPTIONS = {"russia": "ru"}
-IMAGE_BASE = "https://flagpedia.net/data/flags/w580/{}.png"
-
-
-class Country(Converter):
-    """country converter"""
-
-    async def convert(self, ctx: Context, argument: str) -> Union[Dict[str, Union[str, int]], None]:
-        argument = argument.lower()
-        get = lambda **kwargs: pycountry.countries.get(**kwargs)
-
-        if argument in ("england", "scotland", "wales"):
-
-            special_images = {
-                "england": "gb-eng",
-                "wales": "gb-wls",
-                "scotland": "gb-sct"
-            }
-
-            if tabulate:
-                description = box(f"Emoji Information  [:{argument}:]", lang="ini")
-            else:
-                description = box(f"Emoji Information: :{argument}:", lang="yaml")
-
-            emoji = f":{argument}:"
-            country_name = argument.title()
-            image = IMAGE_BASE.format(special_images[argument])
-            name = country_name.title()
-
-            return {
-                "description": description,
-                "emoji": emoji,
-                "name": square(country_name.title()),
-                "title": f"{emoji} {name}",
-                "image": image
-            }
-
-        else:
-
-            obj = get(name=argument) or get(alpha_2=argument)
-            if not obj:
-                obj = None
-                for k, v in EXCEPTIONS.items():
-                    if k in argument:
-                        obj = get(alpha_2=v)
-                        break
-                if not obj:
-                    raise BadArgument("Could not match this argument to a country.")
-            
-            ret = {
-                "name": square(obj.name.title()), 
-                "title": f":flag_{obj.alpha_2.lower()}: {obj.name}",
-                "emoji": square(f":flag_{obj.alpha_2.lower()}:"),
-                "image": IMAGE_BASE.format(obj.alpha_2.lower())
-            }
-
-            for attr in ("alpha_2", "alpha_3", "numeric", "official_name"):
-                if hasattr(obj, attr):
-                    ret[attr] = square(getattr(obj, attr))
-                
-            return ret
 
 class Flags(Cog):
     """Get flags from country names."""
@@ -118,7 +48,7 @@ class Flags(Cog):
 
 
     @commands.command()
-    async def flag(self, ctx: Context, *, argument: Country):
+    async def flag(self, ctx: Context, *, argument: CountryConverter):
         """Get the flag for a country.
         
         Either the country name or alpha 2 code can be provided.
@@ -142,7 +72,6 @@ class Flags(Cog):
                 description = box(tabulate.tabulate([(format_attr(k), v) for k, v in argument.items()], tablefmt="plain"), lang="ini")
             else:
                 description = box((f"{format_attr(k)}: {v}" for k, v in argument.items()), lang="yaml")
-
 
 
         embed = discord.Embed(
