+++ flags/functions.py
@@ -1,0 +1,10 @@
+def square(t):
+    return f"[{t}]"
+
+
+def format_attr(t):
+    return t.replace('_',' ').title()
+
+
+EXCEPTIONS = {"russia": "ru"}
+IMAGE_BASE = "https://flagpedia.net/data/flags/w580/{}.png"
